working sampling fonts; Coll2D impl

+++ Engine/Rendering/UI/FontReader/Character.cs
@@ -1,17 +1,17 @@
 ﻿namespace Dwarf.Engine.Rendering.UI.FontReader;
 public class Character {
   private int _id;
-  private double _xTextureCoord;
-  private double _yTextureCoord;
-  private double _xMaxTextureCoord;
-  private double _yMaxTextureCoord;
-  private double _xOffset;
-  private double _yOffset;
-  private double _sizeX;
-  private double _sizeY;
-  private double _xAdvance;
+  private float _xTextureCoord;
+  private float _yTextureCoord;
+  private float _xMaxTextureCoord;
+  private float _yMaxTextureCoord;
+  private float _xOffset;
+  private float _yOffset;
+  private float _sizeX;
+  private float _sizeY;
+  private float _xAdvance;
 
-  public Character(int id, double xTextureCoord, double yTextureCoord, double xMaxTextureCoord, double yMaxTextureCoord, double xOffset, double yOffset, double sizeX, double sizeY, double xAdvance) {
+  public Character(int id, float xTextureCoord, float yTextureCoord, float xMaxTextureCoord, float yMaxTextureCoord, float xOffset, float yOffset, float sizeX, float sizeY, float xAdvance) {
     _id = id;
     _xTextureCoord = xTextureCoord;
     _yTextureCoord = yTextureCoord;
@@ -25,13 +25,13 @@ public class Character {
   }
 
   public int Id => _id;
-  public double XTextureCoord => _xTextureCoord;
-  public double YTextureCoord => _yTextureCoord;
-  public double XMaxTextureCoord => _xMaxTextureCoord;
-  public double YMaxTextureCoord => _yMaxTextureCoord;
-  public double XOffset => _xOffset;
-  public double YOffset => _yOffset;
-  public double SizeX => _sizeX;
-  public double SizeY => _sizeY;
-  public double XAdvance => _xAdvance;
+  public float XTextureCoord => _xTextureCoord;
+  public float YTextureCoord => _yTextureCoord;
+  public float XMaxTextureCoord => _xMaxTextureCoord;
+  public float YMaxTextureCoord => _yMaxTextureCoord;
+  public float XOffset => _xOffset;
+  public float YOffset => _yOffset;
+  public float SizeX => _sizeX;
+  public float SizeY => _sizeY;
+  public float XAdvance => _xAdvance;
 }
