+++ Engine/Rendering/UI/FontReader/TextMeshCreator.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Dwarf.Engine.Rendering.UI.FontReader;
 public class TextMeshCreator {
-  public const double LineHeight = 0.03f;
+  public const float LineHeight = 0.03f;
   public const int SpaceAscii = 32;
 
   private FontFile _metaData;
